requirements.txt
google-generativeai
streamlit
python-dotenv
PyPDF2
langchain
langchain_google_genai